+++ raphtory/src/python/graph/utils.rs
@@ -2,7 +2,7 @@ use crate::core::{entities::graph::tgraph::InternalGraph, utils::errors::GraphEr
 use pyo3::{prelude::*, types::IntoPyDict};
 use std::collections::HashMap;
 
-use super::pandas::{
+use super::io::{
     dataframe::{process_pandas_py_df, GraphLoadException},
     loaders::{
         load_edges_from_df, load_edges_props_from_df, load_node_props_from_df, load_nodes_from_df,
